re-organize types into logical files
